perf: volsurface double speed up, rust f64->f32 to cuda

+++ src/main/java/com/erayt/cuda/GpuAlgorithm.java
@@ -3,5 +3,5 @@ package com.erayt.cuda;
 public interface GpuAlgorithm {
     int getId(); // 每个算法唯一标识
 
-    float[] toArgs(); // 转换为 GPU 调用参数
+    double[] toArgs(); // 转换为 GPU 调用参数
 }
